requirements.txt
django>=4.0.0,<4.1.0
djangorestframework~=3.12.2
pyyaml
requests
django-cors-headers